Lots of rearranging. Controllers/routes/mappers(soon) moved to api folder for clutter control. Models fixed for proper ERD and data flow - Matrix can have multiple Samples, Sample can have multiple Results, Matrix can have multiple Methods, Methods can have multiple Analytes, MethodAnalyte table can have multiple Methods and Analyte combinations, MethodAnalyte can have multiple Results. Seeders fixed for testing with real(ish) data, fixed data gen to use real UUIDs for Sample and Result creation (needs Matrix for Sample, and MethodAnalyte for Result based on Matrix of Sample). Insomnia route testing returning good data, BE is looking a lot cleaner. Next step is going to be data mapper for API returns and FE API layer before moving to the FE.

+++ server/db/seed/scripts/runSeeds.cjs
@@ -1,25 +1,32 @@
+const arg = process.argv.slice(2);
+const begin = parseInt(arg[0], 10);
+const end = parseInt(arg[1], 10);
 const fs = require("fs");
 const path = require("path");
 const { sequelize } = require("../../index.cjs");
 
-const runSeeds = async () => {
+async function runSeeds(beginIndex, endIndex) {
   try {
-    await sequelize.sync({ force: true });
-    console.log("Tables dropped and recreated");
+    if (beginIndex === 0) {
+      await sequelize.sync({ force: true });
+      console.log("Tables dropped and recreated");
+    }
+
     // sets the absolute path for the current directory that is being manipulated
     const dir = path.join(__dirname, "..", "seeds");
     const files = fs.readdirSync(dir).sort();
-    // loops through the array of files that are in the directory stored here by readdirSync
-    for (const file of files) {
-      if (!file.endsWith(".cjs")) continue;
+    // modally runs either --- 1 or --- 2 which determines if it
+    
+    for (let i = beginIndex; i < endIndex; i++) {
+      if (!files[i].endsWith(".cjs")) continue;
       // the full file from the directory combined
       // loads the current file at the filepath into memory to be ran
-      const seedFunction = require(path.join(dir, file));
+      const seedFunction = require(path.join(dir, files[i]));
       // let us know the seeding is happening on file
-      console.log(`Running Seed: ${file}`);
-      
+      console.log(`Running Seed: ${files[i]}`);
+
       if (typeof seedFunction !== "function") {
-        console.log(`${file} does not export a function`);
+        console.log(`${files[i]} does not export a function`);
       }
       await seedFunction();
     }
@@ -28,6 +35,6 @@ const runSeeds = async () => {
   } catch (err) {
     console.error("Seeding failed: ", err);
   }
-};
+}
 
-runSeeds();
+runSeeds(begin, end);
